+++ src/test/java/nextstep/subway/acceptance/AuthAcceptanceTest.java
@@ -2,15 +2,16 @@ package nextstep.subway.acceptance;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.utils.DataLoader;
 import nextstep.subway.stub.GithubResponses;
+import nextstep.subway.utils.DataLoader;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import static nextstep.subway.acceptance.LoginSteps.깃헙_로그인_요청;
-import static nextstep.subway.acceptance.MemberSteps.베어러_인증_로그인_요청;
+
+import static nextstep.subway.acceptance.LoginSteps.베어러_인증_로그인_요청;
 import static nextstep.subway.acceptance.MemberSteps.토큰으로_내_정보_요청;
 import static org.assertj.core.api.Assertions.assertThat;
