+++ src/test/java/nextstep/subway/acceptance/LineSectionAcceptanceTest.java
@@ -125,17 +125,6 @@ class LineSectionAcceptanceTest extends AcceptanceTest {
         assertThat(response.jsonPath().getList("stations.id", Long.class)).containsExactly(강남역, 정자역);
     }
 
-    private Map<String, String> createLineCreateParams(Long upStationId, Long downStationId) {
-        Map<String, String> lineCreateParams;
-        lineCreateParams = new HashMap<>();
-        lineCreateParams.put("name", "신분당선");
-        lineCreateParams.put("color", "bg-red-600");
-        lineCreateParams.put("upStationId", upStationId + "");
-        lineCreateParams.put("downStationId", downStationId + "");
-        lineCreateParams.put("distance", 10 + "");
-        return lineCreateParams;
-    }
-
     private Map<String, String> createSectionCreateParams(Long upStationId, Long downStationId) {
         Map<String, String> params = new HashMap<>();
         params.put("upStationId", upStationId + "");
