+++ src/test/java/nextstep/subway/acceptance/LoginSteps.java
@@ -24,4 +24,17 @@ public class LoginSteps {
         return response;
     }
 
+    public static ExtractableResponse<Response> 베어러_인증_로그인_요청(String email, String password) {
+        Map<String, String> params = new HashMap<>();
+        params.put("email", email);
+        params.put("password", password);
+
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(params)
+                .when().post("/login/token")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value()).extract();
+    }
+
 }
